[build-script] When we are building a stage 2 stdlib, do not build/perform the stage1 tests.
I also fixed the logic a little bit where even when not building the stage 2
stdlib. Specifically, I changed it so that even in such a case when we are not
building a dynamic stdlib, we do not call the target swift-test-*. This matches
the logic inside CMake around the creation of said targets.