[moveOnly] Implement a generic _move function, a transparent AEIC wrapper around Builtin.move.
Since we do not have opaque values enabled yet, _move must use unconstrained
generics implying that it is dealing with addresses. To work around that, I had
to make a change to Builtin.move: Builtin.move is now a true generic builtin
with a SIL representation. Previously, Builtin.move did not have a SIL
representation and just caused us to emit move_value in SILGen. Since move_value
does not support addresses (and we do not want it to), we instead in SILGen emit
a call to ``builtin "move"`` which does support being called in a generic form.

To make this work with normal code, we take advantage of the fact that the
mandatory inliner specializes as it inlines to enable the inliner to convert as
it inlines:

  builtin "move"
    ->
  load [take] + move_value + store [init]

if the generic type that we are specializing is loadable in our caller. Due to
the way that we SILGen, we know that in _move, the result of calling
Builtin.move() will be in a local alloc_stack in _move and that if we have a
loadable ctx in our caller, then we will have spilled our loadable value into a
local alloc_stack. This is a pattern that predictable dead allocation handles
very well.

To make sure that we do not allow for any weird behavior, we only allow for
builtin "move" to exist within _move (enforced by the SILVerifier checking for
the semantic tag "lifetimemanagement.move" which is only on _move).

To eliminate any potential ABI impact, if someone calls _move in a way that
causes it to be used in a context where the transparent inliner will not inline
it, I taught IRGen that Builtin.move is equivalent to a take from src -> dst and
marked _move as always emit into client (AEIC).