[moveOnly] Add a new instruction called mark_move_addr.
This instruction is similar to a copy_addr except that it marks a move of an
address that has to be checked. In order to keep the memory lifetime verifier
happy, the semantics before the checker runs are the mark_move_addr is
equivalent to copy_addr [init] (not copy_addr [take][init]).

The checker then checks that the mark_move_addr src address isn't used before
being re-initialized again and if it succeeds, converts the instruction to a
copy_addr [take][init]/eliminates any extra destroy_addr. If we fail, we emit a
diagnostic.

NOTE: The move kills checker for addresses is going to run after Mandatory
Inlining, but before predictable memory opts and friends.