[silgen] Teach SIL that ClassMethod can have a @moveOnly as an argument.
I also added a new method to SILType called getSemanticASTType(). The idea is
that one can use this in contexts where one wants to get back either the AST
type or the AST type after removing move only. I changed
SILType::getClassOrGenericClass() to use this since for this query, we want
moveOnly and copyable types to be treated the same.

Additionally, I grepped through the code base for all instances where we used
ClassMethod and I updated any places I saw where we were directly using
getASTType() in a manner that would break for moveOnly classes being passed to
ClassMethod.