[debug-info] Change Swift debug info to use llvm.dbg.addr instead of llvm.dbg.declare.
The reason why I am doing this is that to support the move function, we want to
make it so that in the debugger once a value is moved it is marked as
unavailable. llvm.dbg.declare is unable to support this functionality since
formally it models a value that must be around for the entire call frame. In
contrast, llvm.dbg.addr can have its lifetime ended early by passing undef to
llvm.dbg.value for the same debug metadata.

rdar://85020571