[move-only] Add a new pass called MoveOnlyTypeEliminator that runs after move only checking.
This pass lowers moveonly-ness from the IR after we have finished move only
checking. It does this for both trivial and non-trivial types. For trivial
types, this will remain the behavior forever, but in the future, we will
propagate the moveOnly bit for non-trivial types so that we can catch the
optimizer adding extra copies later in the pipeline. That being said, currently
we do not have this guarantee and this patch at least improves the world and
lets us codegen no implicit copy code again.

The failing test was: noimplicitcopy_attr.swift.