[move-function-addr] Fix a subtle bug where we were not inserting debug info in the case of vars.
The specific problem was that we assume that SILGen will only emit a single
debug_value for a value. This condition breaks once we start processing since we
are inserting extra debug_value when reiniting. The end result is that we do not
insert the undef, addr. To fix this I just hoisted out this computation upon the
address before we do anything.