[move-only] Add a small canonicalization pass that runs before the move checker that fixes small temporary allocations.
Some of these issues I am going to need to actually fix in the move checker
itself, but if I can fix some of these in a separate pass that can be tested
separately, I think it makes sense to do so.

This eliminates a few of the copy of noncopyable type errors around subscripts.