[sending] Replace sending with consuming when suppressing sending from arguments in swiftinterface files.
Previously we would just not print sending. This causes problems since sending
implies a +1 parameter and by removing it we convert the parameter to a +0
parameter, breaking ABI. In this commit, I make it so that when we suppress
sending from argument parameters, we just replace it with consuming so that we
preserve ABI even for callers who do not support sending.

rdar://131066640
(cherry picked from commit fdd435a7827b51bb3d473c9ff4eb32eeb086f997)